requirements.txt
discord.py
requests
solders
stegano